Separate vanilla JS functions into their own files
