+++ viser/client/src/Utils.tsx
@@ -1,0 +1,7 @@
+import { GuiConfig } from "./ControlPanel/GuiState";
+import * as Messages from "./WebsocketMessages";
+
+// GUIState Type Guard
+export function isGuiConfig(message: Messages.Message): message is GuiConfig {
+    return message.type.startsWith("GuiAdd");
+}
