+++ viser/client/src/WebsocketFunctions.tsx
@@ -1,0 +1,52 @@
+import { MutableRefObject } from "react";
+import { Message } from "./WebsocketMessages";
+import { pack } from "msgpackr";
+
+/** Send message over websocket. */
+export function sendWebsocketMessage(
+    websocketRef: MutableRefObject<WebSocket | null>,
+    message: Message
+) {
+    if (websocketRef.current === null) return;
+    websocketRef.current.send(pack(message));
+}
+
+/** Returns a function for sending messages, with automatic throttling. */
+export function makeThrottledMessageSender(
+    websocketRef: MutableRefObject<WebSocket | null>,
+    throttleMilliseconds: number
+) {
+    let readyToSend = true;
+    let stale = false;
+    let latestMessage: Message | null = null;
+
+    function send(message: Message) {
+        if (websocketRef.current === null) return;
+        latestMessage = message;
+        if (readyToSend) {
+            websocketRef.current.send(pack(message));
+            stale = false;
+            readyToSend = false;
+
+            setTimeout(() => {
+                readyToSend = true;
+                if (!stale) return;
+                latestMessage && send(latestMessage);
+            }, throttleMilliseconds);
+        } else {
+            stale = true;
+        }
+    }
+    return send;
+}
+
+/** Type guard for threejs textures. Meant to be used with `scene.background`. */
+export function isTexture(
+    background: THREE.Color | THREE.Texture | THREE.CubeTexture | null | undefined
+): background is THREE.Texture {
+    return (
+        background !== null &&
+        background !== undefined &&
+        (background as THREE.Texture).isTexture !== undefined
+    );
+}
