+++ viser/client/src/WebsocketInterface.tsx
@@ -1,67 +1,19 @@
 import AwaitLock from "await-lock";
-import { pack, unpack } from "msgpackr";
+import { unpack } from "msgpackr";
 
-import React, { MutableRefObject, useContext } from "react";
+import React, { useContext } from "react";
 import * as THREE from "three";
 import { TextureLoader } from "three";
 
-import { ViewerContext } from ".";
-import { isGuiConfig } from "./ControlPanel/GuiState";
+import { ViewerContext } from "./App";
+import { isGuiConfig } from "./Utils";
 import { SceneNode } from "./SceneTree";
 import { syncSearchParamServer } from "./SearchParamsUtils";
 import { CameraFrustum, CoordinateFrame } from "./ThreeAssets";
 import { Message } from "./WebsocketMessages";
 import styled from "@emotion/styled";
 import { Html, PivotControls } from "@react-three/drei";
-
-/** Send message over websocket. */
-export function sendWebsocketMessage(
-  websocketRef: MutableRefObject<WebSocket | null>,
-  message: Message
-) {
-  if (websocketRef.current === null) return;
-  websocketRef.current.send(pack(message));
-}
-
-/** Returns a function for sending messages, with automatic throttling. */
-export function makeThrottledMessageSender(
-  websocketRef: MutableRefObject<WebSocket | null>,
-  throttleMilliseconds: number
-) {
-  let readyToSend = true;
-  let stale = false;
-  let latestMessage: Message | null = null;
-
-  function send(message: Message) {
-    if (websocketRef.current === null) return;
-    latestMessage = message;
-    if (readyToSend) {
-      websocketRef.current.send(pack(message));
-      stale = false;
-      readyToSend = false;
-
-      setTimeout(() => {
-        readyToSend = true;
-        if (!stale) return;
-        latestMessage && send(latestMessage);
-      }, throttleMilliseconds);
-    } else {
-      stale = true;
-    }
-  }
-  return send;
-}
-
-/** Type guard for threejs textures. Meant to be used with `scene.background`. */
-export function isTexture(
-  background: THREE.Color | THREE.Texture | THREE.CubeTexture | null | undefined
-): background is THREE.Texture {
-  return (
-    background !== null &&
-    background !== undefined &&
-    (background as THREE.Texture).isTexture !== undefined
-  );
-}
+import { isTexture, makeThrottledMessageSender } from "./WebsocketFunctions";
 
 /** Float **/
 function threeColorBufferFromUint8Buffer(colors: ArrayBuffer) {
